Add ui package, move Activity kotlin files to ui package
