+++ app/src/main/java/com/example/rvbnb/ui/RegisterActivity.kt
@@ -1,8 +1,9 @@
-package com.example.rvbnb
+package com.example.rvbnb.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.rvbnb.R
 import kotlinx.android.synthetic.main.activity_register.*
 
 class RegisterActivity : AppCompatActivity() {
