+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Mon Sep 23 09:59:08 PDT 2019
-sdk.dir=/Users/maitran/Library/Android/sdk
+#Mon Sep 23 13:28:19 MDT 2019
+sdk.dir=C\:\\Users\\beast\\AppData\\Local\\Android\\Sdk
